requirements.txt
amazon-braket-sdk>=1.83.0
boto3>=1.34.0
pycryptodome>=3.20.0
numpy>=1.26.0
